requirements.txt
ijson
lxml
nltk
streamlit
google-genai
python-dotenv